+++ src/main/java/laddergame/domain/ladder/Bridge.java
@@ -1,6 +1,6 @@
 package laddergame.domain.ladder;
 
-import laddergame.domain.player.Position;
+import laddergame.domain.vo.Position;
 import laddergame.domain.vo.Column;
 
 import java.util.Objects;
