+++ src/test/java/laddergame/domain/player/NameTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain.player;
 
+import laddergame.domain.vo.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
