+++ src/test/java/laddergame/domain/player/PositionTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain.player;
 
+import laddergame.domain.vo.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
